Guessing Game in C#
Console App is written in C#
I found a bust version of this online to give me an idea of what it was doing. I looked it over and rewrote it.
The initial writing was to get the basic functionality working.
Then add in some control over what was guessed.
It will only accept numbers between 1 to 100, no other characters
Only allows 5 guesses, error guesses did not increment the guess count
Clears the console at the start of the game and just before the end, if the game is won.