List of school kid sums in string
Given a list of sums that a school kid would get. Each sum is a string.
Calculate the answer of each with reusable code.

The original test (part of a freecodecamp Q) was to format the sum vertical from the string, not to compute the answer.  This just computes the answer and consoles out the result. This could be used to enhance the original question.